move filter options type to better location
