+++ frontend/src/components/search/SearchFilterAccordion/SearchFilterOptions.ts
@@ -1,7 +1,6 @@
 import { omit } from "lodash";
 import { eligibilityTypes } from "src/constants/opportunity";
-
-import { FilterOption } from "src/components/search/SearchFilterAccordion/SearchFilterAccordion";
+import { FilterOption } from "src/types/search/searchResponseTypes";
 
 export const eligibilityOptions: FilterOption[] = eligibilityTypes.map((type) =>
   omit(type, "group"),
