+++ frontend/tests/components/search/SearchFilterAccordion/SearchFilterSection/SectionLinkLabel.test.tsx
@@ -1,11 +1,11 @@
 import "@testing-library/jest-dom";
 
 import { axe } from "jest-axe";
+import { FilterOption } from "src/types/search/searchResponseTypes";
 import { render, screen } from "tests/react-utils";
 
 import React from "react";
 
-import { FilterOption } from "src/components/search/SearchFilterAccordion/SearchFilterAccordion";
 import SectionLinkLabel from "src/components/search/SearchFilterAccordion/SearchFilterSection/SectionLinkLabel";
 
 // Mock the Icon component from "@trussworks/react-uswds"
